requirements.txt
azure-identity>=1.15.0
pyodbc>=5.0.0
openpyxl>=3.1.0
pymupdf>=1.23.0
python-pptx>=0.6.23
python-docx>=1.1.0
